added the ability to find spikes in C++ code
this can massively reduce the memory footprint of very long
simulations. the way this can help is you can now run your
simulations with a very large dt, and choose output_type =2
even though this records all the variables, the huge
savings on the time resolution can mean that your
memory use can be much lower than if you just pull out V.
spiketimes are computed with a precision of sim_dt, so
there is no penalty on picking dt as large as you want.